requirements.txt
bigartm==0.9.2
nltk==3.6.2
pymorphy2==0.9.1